Replace Shout with a new module SecureShell
This change fixes a few issues in the project:

1. Shout has some odd compliation issues as noted in Shout#34
2. Shout is based on libssh2 and libssh2 does not support certificate
based authentication.
3. Shout did not capture standard error from remote.

This patch creates 2 new modules. Clibssh and SecureShell. Clibssh is a
system package that imports libssh into Swift. Clibssh is not expected
to be used directly. SecureShell wraps Clibssh and provides the API that
is expected to be used by gitlab-fusion.